+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/aio/_document_model_administration_client_async.py
@@ -414,7 +414,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         if self._api_version == DocumentAnalysisApiVersion.V2022_08_31:
             _client_op_path = self._client.get_resource_details
         response = await _client_op_path(**kwargs)
-        return ResourceDetails._from_generated(response.custom_document_models)
+        return ResourceDetails._from_generated(response)
 
     @distributed_trace_async
     async def get_document_model(self, model_id: str, **kwargs: Any) -> DocumentModelDetails: