+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_dmac_mgmt.py
@@ -85,6 +85,9 @@ class TestManagement(FormRecognizerTest):
 
         assert info.custom_document_models.limit
         assert info.custom_document_models.count
+        assert info.custom_neural_document_model_builds.quota
+        assert info.custom_neural_document_model_builds.quota_reset_date_time
+        assert info.custom_neural_document_model_builds.used is not None
 
     @skip_flaky_test
     @FormRecognizerPreparer()