+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `features` keyword argument on `begin_analyze_document()` and `begin_analyze_document_from_url()`.
 - Added `AnalysisFeature` enum with optional document analysis feature to enable.
 - Added the following optional properties on `DocumentStyle` class: `similar_font_family`, `font_style`, `font_weight`, `color`, `background_color`.
+- Added model `QuotaDetails` and property `custom_neural_document_model_builds` on `ResourceDetails`
 
 ### Breaking Changes